+++ tests/sync/test_embed.py
@@ -1,10 +1,16 @@
 import random
 import string
 import unittest
+from typing import List
 
 from utils import get_api_key
 
 import cohere
+from cohere.responses.embeddings import (
+    EMBEDDINGS_BY_TYPE_RESPONSE_TYPE,
+    EMBEDDINGS_FLOATS_RESPONSE_TYPE,
+    EmbeddingsByType,
+)
 
 API_KEY = get_api_key()
 co = cohere.Client(API_KEY)
@@ -102,3 +108,25 @@ class TestEmbed(unittest.TestCase):
         for predictionExpected, predictionActual in zip(predictionsExpected, list(predictionsActual)):
             for elementExpected, elementAcutal in zip(predictionExpected, predictionActual):
                 self.assertAlmostEqual(elementExpected, elementAcutal, places=1)
+
+    def test_success_embeddings_floats_response_type(self):
+        prediction = co.embed(
+            model="small",
+            texts=["python", "golang", "typescript"],
+        )
+        self.assertEqual(prediction.response_type, EMBEDDINGS_FLOATS_RESPONSE_TYPE)
+        assert isinstance(prediction.embeddings, List)
+
+    def test_success_embeddings_by_type_response_type(self):
+        prediction = co.embed(
+            model="small",
+            texts=["python", "golang", "typescript"],
+            embedding_types=["float"],
+        )
+        self.assertEqual(prediction.response_type, EMBEDDINGS_BY_TYPE_RESPONSE_TYPE)
+        assert isinstance(prediction.embeddings, EmbeddingsByType)
+        assert prediction.embeddings.float
+        assert not prediction.embeddings.int8
+        assert not prediction.embeddings.uint8
+        assert not prediction.embeddings.binary
+        assert not prediction.embeddings.ubinary